+++ app/routes/components/Typography/changelog.codegen.ts
@@ -14,4 +14,5 @@ export const changelog = ''
   + '- 2025-10-12: Add `Typography.Root` component as a wrapper for all typography\n'
   + '  elements. Improve overall styles and spacing. Put styles into `components` CSS\n'
   + '  layer.\n'
+  + '- 2025-10-13: Add `Typography.CodeBlock` component.\n'
   ;