+++ src/components/Typography/index.ts
@@ -1,3 +1,3 @@
-// Typography from copy-ui @ 2025-10-12.
+// Typography from copy-ui @ 2025-10-13.
 
 export { Typography } from "./Typography";